+++ src/Manager/RequestManager/RequestManagerInterface.php
@@ -32,9 +32,6 @@ interface RequestManagerInterface {
   const ERROR_USER_INVITE = 4;
   const ERROR_GET_USERS = 5;
   const ERROR_SET_USER_PASS = 6;
-  const ERROR_GET_FILE_INFO = 7;
-  const ERROR_GET_FOLDER_INFO = 8;
-  const ERROR_SHARE_ROOM = 9;
 
   /**
    * Connect to ONLYOFFICE DocSapace.
@@ -84,16 +81,6 @@ interface RequestManagerInterface {
   public function inviteToDocSpace($email, $password_hash, $firstname, $lastname, $type, $token = NULL);
 
   /**
-   * Create ONLYOFFICE DocSpace public user.
-   *
-   * @param string $url
-   *   The ONLYOFFICE DocSpace URL.
-   * @param string $token
-   *   The ONLYOFFICE DocSpace token.
-   */
-  public function createPublicUser($url, $token);
-
-  /**
    * Set ONLYOFFICE DocSpace user password.
    *
    * @param string $user_id
@@ -104,29 +91,5 @@ interface RequestManagerInterface {
    *   The ONLYOFFICE DocSpace token.
    */
   public function setUserPassword($user_id, $password_hash, $token);
-
-  /**
-   * Get file information from ONLYOFFICE DocSpace.
-   *
-   * @param string $file_id
-   *   The ONLYOFFICE DocSpace file ID.
-   */
-  public function getFileInfo($file_id);
-
-  /**
-   * Get folder information from ONLYOFFICE DocSpace.
-   *
-   * @param string $folder_id
-   *   The ONLYOFFICE DocSpace folder ID.
-   */
-  public function getFolderInfo($folder_id);
-
-  /**
-   * Share room public user ONLYOFFICE DocSpace.
-   *
-   * @param string $room_id
-   *   The ONLYOFFICE DocSpace room ID.
-   */
-  public function shareRoomPublicUser($room_id);
 
 }