+++ src/Manager/UtilsManager/UtilsManager.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\onlyoffice_docspace\Manager\ComponentManager;
+namespace Drupal\onlyoffice_docspace\Manager\UtilsManager;
 
 /**
  * Copyright (c) Ascensio System SIA 2023.
@@ -26,13 +26,12 @@ use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Path\CurrentPathStack;
 use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\Url;
-use Drupal\onlyoffice_docspace\Controller\OODSPCredentialsController;
 use Drupal\onlyoffice_docspace\Manager\ManagerBase;
 
 /**
- * The ONLYOFFICE DocSpace Component Manager.
+ * The ONLYOFFICE DocSpace Utils Manager.
  */
-class ComponentManager extends ManagerBase {
+class UtilsManager extends ManagerBase {
 
   /**
    * Locales for ONLYOFFICE DocSpace.
@@ -99,50 +98,61 @@ class ComponentManager extends ManagerBase {
    */
   public function buildComponent(array $build, AccountInterface $user) {
     $isAnonymous = $user->isAnonymous();
-    $email = $isAnonymous ? OODSPCredentialsController::OODSP_PUBLIC_USER_LOGIN : $user->getEmail();
+    $currentUser = $user->getEmail();
 
-    if ($user->hasPermission('administer onlyoffice_docspace configuration')) {
-      $error_message = $this->t('Go to the settings to configure ONLYOFFICE DocSpace connector.');
-    }
-    else {
-      $error_message = $this->t('Portal unavailable! Please contact the administrator!');
-    }
-
-    $unauthorized_header = $this->t('Authorization unsuccessful');
-    $unauthorized_message = $this->t('Please contact the administrator.');
+    $messageDocspaceUnavailable = $this->t('Portal unavailable! Please contact the administrator!');
 
     if (!$isAnonymous) {
-      $unauthorized_message = $this->t(
-        'Please go to <a href="@login_page">ONLYOFFICE DocSpace Login page</a> and enter your password to restore access.',
-        [
-          '@login_page' => Url::fromRoute(
-            'onlyoffice_docspace.page_login',
-            [],
-            ['query' => ['redirect' => $this->currentPath->getPath()]]
-          )->toString(),
-        ]
-      );
+      if ($user->hasPermission('administer onlyoffice_docspace configuration')) {
+        $messageDocspaceUnavailable = $this->t('Go to the settings to configure ONLYOFFICE DocSpace connector.');
+      }
+
+      $messageUnauthorizedHeader = $this->t('Authorization unsuccessful');
+      $messageUnauthorizedMessage = $this->t('Please contact the administrator.');
+
+      if ($user->hasPermission('administer site configuration')) {
+        $messageUnauthorizedMessage = $this->t(
+          'Please go to <a href="@login_page">ONLYOFFICE DocSpace Login page</a> and enter your password to restore access.',
+          [
+            '@login_page' => Url::fromRoute(
+              'onlyoffice_docspace.page_login',
+              [],
+              ['query' => ['redirect' => $this->currentPath->getPath()]]
+            )->toString(),
+          ]
+        );
+      }
+    }
+    else {
+      $messageUnauthorizedHeader = $this->t('Access denied!');
+      $messageUnauthorizedMessage = $this->t('Please log in to the site!');
     }
 
-    $build['#attached']['library'][] = 'onlyoffice_docspace/onlyoffice_docspace.component';
+    $build['#attached']['library'][] = 'onlyoffice_docspace/onlyoffice_docspace.docspace-integration-sdk';
+    $build['#attached']['library'][] = 'onlyoffice_docspace/onlyoffice_docspace.utils';
 
-    $build['#attached']['drupalSettings']['DocSpaceComponent'] = [
-      'currentUser' => $email,
+    $build['#attached']['drupalSettings']['OODSP_Settings'] = [
+      'currentUser' => $currentUser,
       'isPublic' => $isAnonymous,
+      'isAnonymous' => $isAnonymous,
       'url' => rtrim($this->config('onlyoffice_docspace.settings')->get('url'), "/") . '/',
       'ajaxUrl' => Url::fromRoute('onlyoffice_docspace.credentilas')->setAbsolute()->toString(),
       'loginUrl' => Url::fromRoute('onlyoffice_docspace.page_login')->setAbsolute()->toString(),
       'adminUrl' => Url::fromRoute('onlyoffice_docspace.page')->setAbsolute()->toString(),
       'messages' => [
-        'error' => $error_message,
-        'unauthorized_header' => $unauthorized_header,
-        'unauthorized_message' => $unauthorized_message,
+        'docspaceUnavailable' => $messageDocspaceUnavailable,
+        'unauthorizedHeader' => $messageUnauthorizedHeader,
+        'unauthorizedMessage' => $messageUnauthorizedMessage,
       ],
       'images' => [
-        'onlyoffice'  => '/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/onlyoffice.svg',
+        'onlyoffice'  => '/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/onlyoffice-docspace.svg',
         'unavailable' => '/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/unavailable.svg',
         'room-icon'   => '/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/room-icon.svg',
         'file-icon'   => '/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/file-icon.svg',
+      ],
+      'labels' => [
+        'room' => $this->t('DocSapce Room'),
+        'file' => $this->t('DocSapce File'),
       ],
       'locale' => $this->getLocaleForDocspace(),
     ];