+++ src/Plugin/Field/FieldWidget/OODSPWidget.php
@@ -28,8 +28,7 @@ use Drupal\Core\Field\FieldItemListInterface;
 use Drupal\Core\Field\WidgetBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\Core\StringTranslation\TranslatableMarkup;
-use Drupal\onlyoffice_docspace\Manager\ComponentManager\ComponentManager;
+use Drupal\onlyoffice_docspace\Manager\UtilsManager\UtilsManager;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
@@ -46,11 +45,11 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class OODSPWidget extends WidgetBase {
 
   /**
-   * The ONLYOFFICE DocSpace Component manager.
+   * The ONLYOFFICE DocSpace Utils manager.
    *
-   * @var \Drupal\onlyoffice_docspace\Manager\ComponentManager\ComponentManager
+   * @var \Drupal\onlyoffice_docspace\Manager\UtilsManager\UtilsManager
    */
-  protected $componentManager;
+  protected $utilsManager;
 
   /**
    * Current user service.
@@ -86,8 +85,8 @@ class OODSPWidget extends WidgetBase {
    *   The widget settings.
    * @param array $third_party_settings
    *   Any third party settings.
-   * @param \Drupal\onlyoffice_docspace\Manager\ComponentManager\ComponentManager $component_manager
-   *   The ONLYOFFICE DocSpace Request manager.
+   * @param \Drupal\onlyoffice_docspace\Manager\UtilsManager\UtilsManager $utils_manager
+   *   The ONLYOFFICE DocSpace Utils manager.
    * @param \Drupal\Core\Session\AccountInterface $current_user
    *   Current user service.
    * @param \Drupal\Core\Config\ConfigFactoryInterface $config_factory
@@ -101,13 +100,13 @@ class OODSPWidget extends WidgetBase {
     FieldDefinitionInterface $field_definition,
     array $settings,
     array $third_party_settings,
-    ComponentManager $component_manager,
+    UtilsManager $utils_manager,
     AccountInterface $current_user,
     ConfigFactoryInterface $config_factory,
     ModuleExtensionList $extension_list_module
   ) {
     parent::__construct($plugin_id, $plugin_definition, $field_definition, $settings, $third_party_settings);
-    $this->componentManager = $component_manager;
+    $this->utilsManager = $utils_manager;
     $this->currentUser = $current_user;
     $this->configFactory = $config_factory;
     $this->extensionListModule = $extension_list_module;
@@ -123,7 +122,7 @@ class OODSPWidget extends WidgetBase {
       $configuration['field_definition'],
       $configuration['settings'],
       $configuration['third_party_settings'],
-      $container->get('onlyoffice_docspace.component_manager'),
+      $container->get('onlyoffice_docspace.utils_manager'),
       $container->get('current_user'),
       $container->get('config.factory'),
       $container->get('extension.list.module'),
@@ -139,7 +138,7 @@ class OODSPWidget extends WidgetBase {
       'class' => ['onlyoffice-docspace-widget'],
     ];
 
-    $element = $this->componentManager->buildComponent($element, $this->currentUser);
+    $element = $this->utilsManager->buildComponent($element, $this->currentUser);
     $element['#attached']['library'][] = 'onlyoffice_docspace/onlyoffice_docspace.widget';
 
     $element['target_id'] = [
@@ -161,6 +160,15 @@ class OODSPWidget extends WidgetBase {
       '#type' => 'hidden',
       '#default_value' => $items[$delta]->image ?? '',
     ];
+
+    $element['request_token'] = [
+      '#type' => 'hidden',
+      '#default_value' => $items[$delta]->request_token ?? '',
+    ];
+
+    if (empty($items[$delta]->request_token)) {
+      $publicIndex = 'hidden';
+    }
 
     $element['fields'] = [
       '#type' => 'html_tag',
@@ -187,19 +195,17 @@ class OODSPWidget extends WidgetBase {
           '#type' => 'html_tag',
           '#tag' => 'div',
           '#attributes' => [
-            'class' => ['oodsp-container-items'],
+            'class' => ['oodsp-container-items', 'form-item__label'],
+          ],
+          'type' => [
+            '#type' => 'html_tag',
+            '#tag' => 'div',
+            '#value' => '<p><span class="value">' . $this->getTitleType($items[$delta]->type) . '</span><img class="public-index ' . $publicIndex . '" src="/' . $this->extensionListModule->getPath('onlyoffice_docspace') . '/images/public.svg" /></p>',
           ],
           'title' => [
-            '#type' => 'textfield',
-            '#title' => new TranslatableMarkup('Title'),
-            '#default_value' => $items[$delta]->title ?? '',
-            '#disabled' => TRUE,
-            '#maxlength' => 1024,
-            '#weight' => -11,
-            '#wrapper_attributes' => [
-              'id' => 'title',
-              'class' => ['oodsp-container-inline'],
-            ],
+            '#type' => 'html_tag',
+            '#tag' => 'div',
+            '#value' => '<p>' . $this->getTranslateTilte() . ': <span class="value font-weight-normal">' . ($items[$delta]->title ?? '') . '</span></p>',
           ],
         ],
       ],
@@ -271,6 +277,28 @@ class OODSPWidget extends WidgetBase {
     ];
 
     return $images[$type];
+  }
+
+  /**
+   * Returns title for type entity.
+   *
+   * @param string $type
+   *   The item type.
+   */
+  private function getTitleType($type) {
+    $images = [
+      'manager' => $this->t('DocSpace Room'),
+      'editor' => $this->t('DocSpace File'),
+    ];
+
+    return $images[$type];
+  }
+
+  /**
+   * Returns translate.
+   */
+  private function getTranslateTilte() {
+    return $this->t('Title');
   }
 
 }