+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 ## 1.0.0
 ## Added
+- support for public rooms